#39: Rename root module
