+++ doc/changes/changes_0.1.0.md
@@ -25,7 +25,7 @@ Code name: Initial Release
 * #18: Replaced lunajson with cjson
 * #17: Added unit tests
 * #35: Added integration test for fetching large results with fetch size smaller than the row size
-* #33: Replaced `cjson.null` in returned row data with `luasqlexasol.NULL`
+* #33: Replaced `cjson.null` in returned row data with `luasql.exasol.NULL`
 
 ## Documentation
