+++ exasol-driver-lua-0.1.0-1.rockspec
@@ -53,18 +53,18 @@ test = {
 build = {
     type = "builtin",
     modules = {
-        luasqlexasol = "src/luasqlexasol.lua",
-        ConnectionProperties = "src/ConnectionProperties.lua",
-        Connection = "src/Connection.lua",
-        constants = "src/constants.lua",
-        CursorData = "src/CursorData.lua",
-        Cursor = "src/Cursor.lua",
-        Environment = "src/Environment.lua",
-        ExasolWebsocket = "src/ExasolWebsocket.lua",
-        luws = "src/luws.lua",
-        util = "src/util.lua",
-        WebsocketDatahandler = "src/WebsocketDatahandler.lua",
-        Websocket = "src/Websocket.lua",
+        ["luasql.exasol"] = "src/luasql/exasol.lua",
+        ["luasql.exasol.ConnectionProperties"] = "src/luasql/exasol/ConnectionProperties.lua",
+        ["luasql.exasol.Connection"] = "src/luasql/exasol/Connection.lua",
+        ["luasql.exasol.constants"] = "src/luasql/exasol/constants.lua",
+        ["luasql.exasol.CursorData"] = "src/luasql/exasol/CursorData.lua",
+        ["luasql.exasol.Cursor"] = "src/luasql/exasol/Cursor.lua",
+        ["luasql.exasol.Environment"] = "src/luasql/exasol/Environment.lua",
+        ["luasql.exasol.ExasolWebsocket"] = "src/luasql/exasol/ExasolWebsocket.lua",
+        ["luasql.exasol.luws"] = "src/luasql/exasol/luws.lua",
+        ["luasql.exasol.util"] = "src/luasql/exasol/util.lua",
+        ["luasql.exasol.WebsocketDatahandler"] = "src/luasql/exasol/WebsocketDatahandler.lua",
+        ["luasql.exasol.Websocket"] = "src/luasql/exasol/Websocket.lua",
     },
     copy_directories = { "doc" }
 }
