+++ spec/unit/Connection_spec.lua
@@ -1,8 +1,8 @@
 ---@diagnostic disable: undefined-global
 -- luacheck: globals describe it before_each after_each
 require("busted.runner")()
-local connection = require("Connection")
-local ConnectionProperties = require("ConnectionProperties")
+local connection = require("luasql.exasol.Connection")
+local ConnectionProperties = require("luasql.exasol.ConnectionProperties")
 local resultstub = require("resultstub")
 local config = require("config")
 config.configure_logging()
