+++ spec/unit/constants_spec.lua
@@ -1,7 +1,7 @@
 ---@diagnostic disable: undefined-global
 -- luacheck: globals describe it before_each after_each
 require("busted.runner")()
-local constants = require("constants")
+local constants = require("luasql.exasol.constants")
 local config = require("config")
 config.configure_logging()
