+++ src/luasql/exasol/ConnectionProperties.lua
