+++ src/luasql/exasol/constants.lua
@@ -2,7 +2,7 @@
 -- @module constants
 local constants = {}
 
-local util = require("util")
+local util = require("luasql.exasol.util")
 
 --- The version of this module
 constants.VERSION = "0.1.0"
