+++ src/luasql/exasol/util.lua
